Added Particle Effects
I have added particle effects to the pick up and made the lotus images
smaller in paint (because I don't know how to make it smaller in Unity
for now). Also I have edited the lotus code a bit so that it gives the
correct amount of chi or health back to the player.

+++ Assets/Script/HUD/NormalGame/Health.cs
@@ -41,8 +41,9 @@ public class Health : MonoBehaviour {
 		healthStat.setCurrentHealth (healthStat.getTotalHealth());
 	}
 
-	public void RestoreHealth(int amount)
+	public void RestoreHealth(float amount)
 	{
-		healthStat.increaseCurrentHealth (amount);
+
+		healthStat.increaseCurrentHealth ((int)amount);
 	}
 }
